Change architecture (merge all commands for polls to\questions & ... to \polls)
