+++ tgbot/handlers/help.py
@@ -6,12 +6,9 @@ router = Router()
 help_text = """🤖 Список команд:\n\
 \n/start - Перезапустить бота.\
 \n/help - Справка о командах.\
-\n/create_question - Создание вопроса.\
-\n/questions_list - Список созданных опросов.\
-\n/create_poll - Создать опрос с вопросом. \
-\n/polls_list - Список опросов. \
-\n/connect_poll - Подключиться к созданному опросу. \
-\n/show_answers - Список ответов из опросов. \
+\n/questions - Вопросы.\
+\n/polls - Опросы.  \
+\n/answers - Список ответов в ваших опросах. \
 """
 
 # /help handler
